pyproject.toml
[project]
name = "ml-demos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gdown>=5.2.0",
    "ipython>=8.29.0",
    "loguru>=0.7.2",
    "matplotlib>=3.9.2",
    "pillow-simd>=9.5.0.post2",
    "pyqt6>=6.7.1",
    "tabulate>=0.9.0",
    "timm==0.6.13",
    "torch>=2.5.1",
    "transformers>=4.46.3",
    "typer>=0.13.1",
]
